Consolidate image constants and unify drop handler logic
