+++ apps/code/src/main/trpc/routers/os.ts
@@ -5,6 +5,7 @@ import type { IAppMeta } from "@posthog/platform/app-meta";
 import type { DialogSeverity, IDialog } from "@posthog/platform/dialog";
 import type { IImageProcessor } from "@posthog/platform/image-processor";
 import type { IUrlLauncher } from "@posthog/platform/url-launcher";
+import { IMAGE_MIME_TYPES } from "@shared/constants/image";
 import { z } from "zod";
 import { container } from "../../di/container";
 import { MAIN_TOKENS } from "../../di/tokens";
@@ -19,19 +20,6 @@ const getDialog = () => container.get<IDialog>(MAIN_TOKENS.Dialog);
 const getAppMeta = () => container.get<IAppMeta>(MAIN_TOKENS.AppMeta);
 const getImageProcessor = () =>
   container.get<IImageProcessor>(MAIN_TOKENS.ImageProcessor);
-
-const IMAGE_MIME_MAP: Record<string, string> = {
-  png: "image/png",
-  jpg: "image/jpeg",
-  jpeg: "image/jpeg",
-  gif: "image/gif",
-  webp: "image/webp",
-  bmp: "image/bmp",
-  ico: "image/x-icon",
-  svg: "image/svg+xml",
-  tiff: "image/tiff",
-  tif: "image/tiff",
-};
 
 const messageBoxOptionsSchema = z.object({
   type: z.enum(["none", "info", "error", "question", "warning"]).optional(),
@@ -305,7 +293,7 @@ export const osRouter = router({
         if (stat.size > input.maxSizeBytes) return null;
 
         const ext = path.extname(input.filePath).toLowerCase().slice(1);
-        const mime = IMAGE_MIME_MAP[ext] ?? "application/octet-stream";
+        const mime = IMAGE_MIME_TYPES[ext] ?? "application/octet-stream";
 
         const buffer = await fsPromises.readFile(input.filePath);
         return `data:${mime};base64,${buffer.toString("base64")}`;
@@ -366,7 +354,7 @@ export const osRouter = router({
     .input(z.object({ filePath: z.string().min(1) }))
     .mutation(async ({ input }) => {
       const ext = path.extname(input.filePath).replace(".", "").toLowerCase();
-      if (!IMAGE_MIME_MAP[ext]) {
+      if (!IMAGE_MIME_TYPES[ext]) {
         throw new Error(`Unsupported image type: .${ext}`);
       }
 
@@ -378,7 +366,7 @@ export const osRouter = router({
       }
 
       const raw = new Uint8Array(await fsPromises.readFile(input.filePath));
-      const inputMime = IMAGE_MIME_MAP[ext];
+      const inputMime = IMAGE_MIME_TYPES[ext];
 
       return downscaleAndPersist(raw, inputMime, path.basename(input.filePath));
     }),
