--- apps/code/src/renderer/features/message-editor/utils/imageUtils.ts
@@ -1,21 +1,0 @@
-const IMAGE_EXTENSIONS = new Set([
-  "png",
-  "jpg",
-  "jpeg",
-  "gif",
-  "webp",
-  "bmp",
-  "ico",
-  "svg",
-  "tiff",
-  "tif",
-]);
-
-export function isImageFile(filename: string): boolean {
-  const ext = filename.split(".").pop()?.toLowerCase();
-  return !!ext && IMAGE_EXTENSIONS.has(ext);
-}
-
-export function isGifFile(filename: string): boolean {
-  return filename.split(".").pop()?.toLowerCase() === "gif";
-}
